textured-triangle: adapted to Magnum changes, fixed, reverted to TGA.
What the hell. The example did not get built at all for quite some time
because of typo in CMakeLists. The example was documenting TGA format,
the error messages told about PNG and the image was actually in JPEG.

Cleaned up and updated the documentation, reverted back to TGA format,
so it is possible to compile and run this example with core Magnum
repository (no external plugins needed).

This reverts commit 70d39cd9b537becb2653c57f085555ad85de8061.