viewer: simplified, removed tangled loader code, handling corner cases.
I'm probably going to abandon the impossibly complex ResourceLoader
classes. The original idea was to provide on-demand, asynchronous
loading, but as the things are, this is impossible to maintain, document
and use in examples without questioning my own sanity.

The application itself knows best what resources are needed and how they
should be loaded, thus there is no need for this on-demand loading. If
the file has some unused portions, they can be detected upfront by just
traversing through the whole data hierarchy, but that's not the point of
this example *at all*. Also, it was "nice" to be prepared for threaded,
asynchronous loading, but there is still nothing like that and it is
practically impossible with OpenGL anyway.

I'm now using AnySceneImporter instead of the ColladaImporter to support
more formats (such as OpenGEX). Some formats also don't have named
meshes/textures/objects, thus it's better to save resources under their
IDs and not under string identifiers (which could be duplicated anyway).

The documentation will be updated later after some more tweaks.