ovr: Properly handle sRGB for mirroring sdk compositor results.
Since the target textures for the HMD have sRGB format now, they would be
blitted onto the mirror texture and then later from the mirror texture onto
the default framebuffer without proper conversion, resulting in different
images in the HMD compared to the mirror.
Therefore mirror texture and default framebuffer need to be in sRGB aswell.

Signed-off-by: Squareys <Squareys@googlemail.com>