CMake: renamed all WITH_* options to WITH_*_EXAMPLE.
If using this repository as a CMake subproject along with Magnum and
others, the options for examples conflict with library options. The
conflict does not result in any error, it just means that you can either
enable both library and the example or disable both of them, but not
enable just one of them. Affected cases:

 * WITH_AUDIO conflicted with WITH_AUDIO for enabling Audio lib in
   Magnum
 * WITH_BULLET conflicted with WITH_BULLET for enabling Bullet
   integration lib in Magnum Integration
 * WITH_OVR conflicted with WITH_OVR for enabling OculusVR integration
   lib in Magnum Integration
 * WITH_PRIMITIVES conflicted with WITH_PRIMITIVES for enabling
   Primitives lib in Magnum
 * WITH_TEXT conflicted with WITH_TEXT for enabling Text lib in Magnum

Updating also all CI build scripts and packages to avoid confusion.
Can't do any backwards compatibility for obvious reasons (it wouldn't
solve anything).