mouseinteraction: WebGL port.
This was fun... until it wasn't. First I thought I could get away with
just blitting the depth buffer when I need it and then reinterpreting the
float value as 32-bit integer and reading that back with R32UI pixel
format. Nope. Chrome worked fine for the first attempt, but Firefox
wasn't. First Firefox said that the blit formats don't match (that was
fortunately fixable with just always requesting 8-bit stencil even though
we don't ever need it so we can then assume Depth24Stencil8 everywhere).
After that it complained I can't just readPixels() with R32UI, so I had to
pack continued fractions of the depth to a RGBA8 output and then recombine
them back. Ugh. And after that it suddenly wasn't preserving the depth
buffer, so I had to blit it in every draw() instead of just on-demand.