picking: make attribute locations match the Generic shaders.
Because normal was bound to location 1 while Shaders::Generic3D::Normal
was location 2, which caused problems when people took the example code
and switched it to use compile() instead.

Actually, it makes no sense to not use compile() here as well, so I just
removed quite a lot of code by doing exactly that.