arcball: make mouse capture work correctly.
Somehow on X11 calling it in the constructor didn't have any effect.
Plus we definitely don't want this to be enabled always as it makes
impossible to close the window or focus something else. This is
currently using a SDL API directly, will update once the wrapper API is
decided on and merged.