textured-triangle-vulkan: implement.
Yeh, sure, just like that, in one longer afternoon. I just read the two
relevant spec sections (samplers and descriptor sets) and then a bit of
the SPIR-V spec and the example code at the start of it to figure out
how exactly to describe the uniform buffer and sampler there.

And then of course wasted three hours on the assembly having

    OpStore %fragmentColor %5

instead of

    OpStore %fragmentColor %8

and giving a washed out pink image instead.

A lot of the example is ugly, especially the RGBA expansion or the
descriptor layout that should have been reflected directly from
the shader instead probably maybe, or the extra copy through a buffer to
upload the image. That's the areas for future work in addition to
wrapping the new APIs.

Also the elephant in the room, which is the whole coordinate system
having origin at top left instead of bottom left but everything
*somehow* still looking alright, probably because both the importer and
exporter Y flip and so it cancels out. Needs fixing.