multidraw: new example.
The original code is from 2021 when the builtin shaders got UBO and
multidraw support, and was extended to expose SSBOs as well. It might
get further extended in the future, but for now this is sufficient I
think.

Unlike most other examples the code has an Emscripten port as well, but
right now it embeds just the silly Suzanne glTF from the viewer example
because I don't have anything better handy. Edit resources.conf to point
it to some better file if you have it.

The screenshot is a LDraw model imported using a very rough
LdrawImporter plugin implementation that doesn't even import materials
correctly and the meshes are randomly broken but the priority is getting
this example out rather than having a polished importer for a toy file
format. Once I sort those out, I'll update the screenshot and add proper
credits, sorry.